refactor(protocols): extract responses tests to crates/protocols/tests/
Move the `#[cfg(test)] mod tests` block out of
`crates/protocols/src/responses.rs` into a standalone integration
test file `crates/protocols/tests/responses.rs`, matching the
convention already used by `background_mode_protocol.rs` and
`skills_protocol.rs`.

What changed
- `crates/protocols/src/responses.rs`: 3660 -> 2396 lines. Removed the
  entire `#[cfg(test)] mod tests { ... }` block (previously L2398..EOF)
  plus the single blank separator line above it.
- `crates/protocols/tests/responses.rs`: new 1262-line integration
  test file. Contains all 41 tests from the removed inline module,
  dedented one level. `use super::*;` was replaced with explicit
  crate-qualified imports:
    * `use openai_protocol::responses::*;` covers the majority of
      types the tests reference (e.g., `ResponseOutputItem`,
      `ResponseInputOutputItem`, `ResponseContentPart`, `ResponseTool`,
      `ResponsesRequest`, `ResponsesToolChoice`, `McpTool`,
      `Annotation`, `BuiltInToolChoiceType`,
      `ImageGenerationCallStatus`, `IncludeField`,
      `RequireApproval{,Mode,Rules,Filter}`,
      `SimpleInputMessageTypeTag`, `SummaryTextContent`,
      `FileDetail`, `StringOrContentParts`, the tool-choice tag
      enums, `ResponsesFunctionToolChoice`, `ToolChoiceOptions`).
    * `use openai_protocol::common::{ContextManagementType, Detail,
      PromptCacheRetention, ToolChoice as ChatToolChoice,
      ToolChoiceValue as ChatToolChoiceValue, ToolReference};`
      resolves the aliases that the original inline test block
      inherited via the private `use super::common::{...}` in
      `responses.rs`. The `ChatToolChoice` / `ChatToolChoiceValue`
      renames are preserved so the test bodies remain byte-for-byte
      identical to the pre-move logic.
    * The one fully-qualified path inside the tests
      (`crate::common::ContextManagementType::Compaction`) was
      shortened to `ContextManagementType::Compaction` now that the
      type is imported.

Why
- The inline `mod tests` block had grown to 1264 lines (34% of
  `responses.rs`). Relocating the tests keeps the production file
  focused on protocol types.
- The protocols crate already uses `crates/protocols/tests/` for
  protocol-surface integration tests (see `background_mode_protocol.rs`
  and `skills_protocol.rs`). Placing the new file here matches that
  convention rather than introducing a sibling source file, which the
  user flagged as incorrect in PR review feedback on the earlier
  attempt.

How
- Integration tests run as a separate binary against the public crate
  API, so no test could rely on `pub(crate)` or `pub(super)` items.
  Verified by checking every symbol the tests reference: all the types
  used inside the test bodies are `pub` items exposed through
  `openai_protocol::responses` and `openai_protocol::common`.
- Only the `use super::*;` line and the outer `mod tests { ... }`
  braces were dropped; test logic is unchanged.
- No source code was edited. `responses.rs` keeps every `pub` item,
  field, variant, and serde attribute untouched.

Verification
- `cargo test -p openai-protocol --test responses` -> 41 passed,
  0 failed, 0 ignored. Matches the pre-move baseline of 41 tests.
- `cargo test -p openai-protocol` -> 53 (lib) + 9 + 41 + 18 + 1 doc
  tests pass, 0 fail. The lib count dropped from 94 to 53 because the
  41 relocated tests now run under the `responses` integration target.
- `cargo fmt --all -- --check` -> clean.
- `cargo clippy -p openai-protocol --tests -- -D warnings` -> clean.

No source-code behavior or public API was modified.
